fix: show custom protocol icon in action picker sidebar
Use the protocol's custom icon (e.g. WETH PNG) in the integration
registry instead of the generic Box icon, so it appears correctly
in the Service dropdown and action picker.
